Allow null date values for From and To dates
When Exporting ADIF data an error occurs when a date value is not selected for both From and To.  The original code looks to be anticipating this, but when either value is null it is not being interpreted correctly and is erroring out as a result.  Line 82 and 87 fail because the string is empty but trying to force a date format.  Proposed change will check for the existance of a value in the $from and $to strings instead.